pyproject.toml
[project]
name = "a-stock-agent"
version = "0.1.0"
description = "ADK Python A-share research analysis agent."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "akshare>=1.16.0",
    "beautifulsoup4>=4.12.0",
    "ebooklib>=0.18",
    "fastapi>=0.115.0",
    "google-adk>=1.0.0",
    "litellm>=1.60.0",
    "numpy>=2.0.0",
    "pandas>=2.2.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "pypdf>=5.0.0",
    "python-multipart>=0.0.20",
    "sentence-transformers>=3.3.0",
    "uvicorn[standard]>=0.34.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
]

[project.scripts]
a-stock-agent-api = "a_stock_agent.api.main:run"
a-stock-agent-import = "a_stock_agent.cli:import_document"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src/a_stock_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"
